Changed 'ssize_t' into 'int' in order to make it compile on windows.

+++ src/import/bmp.cpp
@@ -37,7 +37,7 @@ namespace
 
    void loadBMPbitfields( std::istream& src, size_t w, size_t h, void* ptr, size_t bpp, size_t R, size_t G, size_t B, bool flip )
    {
-      ssize_t ystep = flip ? -3*w : 3*w;      // Add this to the image pointer to skip to the next row
+      int    ystep = flip ? -3*w : 3*w;      // Add this to the image pointer to skip to the next row
       size_t bytePerPixel = bpp/8;
       size_t padding = (w*bytePerPixel) % 2;   // The rows are always word aligned
 
@@ -86,7 +86,7 @@ namespace
 
    void loadBMPcolormap( std::istream& src, size_t w, size_t h, void* ptr, unsigned char* colorMap, bool flip )
    {
-      ssize_t ystep = flip ? -3*w : 3*w;      // Add this to the image pointer to skip to the next row
+      int ystep = flip ? -3*w : 3*w;      // Add this to the image pointer to skip to the next row
 
       unsigned char *cur = (unsigned char*)ptr, *end = (unsigned char*)ptr + h*3*w;
 
@@ -117,7 +117,7 @@ namespace
 
    void loadBMPtruecolor( std::istream& src, size_t w, size_t h, void* ptr, bool flip )
    {
-      ssize_t ystep = flip ? -3*w : 3*w;      // Add this to the image pointer to skip to the next row
+      int ystep = flip ? -3*w : 3*w;      // Add this to the image pointer to skip to the next row
       size_t padding = w*3 % 2;               // The rows are always word aligned
 
       unsigned char *cur = (unsigned char*)ptr, *end = (unsigned char*)ptr + h*3*w;
@@ -201,7 +201,7 @@ CImage::E_LOAD_RESULT CImage::m_loadBmp( std::istream& file )
    //////////////////////// Extract all >interesting< information from the file header ////////////////////////
    size_t  bfOffBits     = ((size_t)header[10]) | ((size_t)header[11])<<8 | ((size_t)header[12])<<16 | ((size_t)header[13])<<24;
    size_t  biWidth       = ((size_t)header[18]) | ((size_t)header[19])<<8 | ((size_t)header[20])<<16 | ((size_t)header[21])<<24;
-   ssize_t biHeight      = ((size_t)header[22]) | ((size_t)header[23])<<8 | ((size_t)header[24])<<16 | ((size_t)header[25])<<24;
+   int     biHeight      = ((size_t)header[22]) | ((size_t)header[23])<<8 | ((size_t)header[24])<<16 | ((size_t)header[25])<<24;
    size_t  biBitCount    = ((size_t)header[28]) | ((size_t)header[29])<<8;
    size_t  biCompression = ((size_t)header[30]) | ((size_t)header[31])<<8 | ((size_t)header[32])<<16 | ((size_t)header[33])<<24;
    size_t  biClrUsed     = ((size_t)header[46]) | ((size_t)header[47])<<8 | ((size_t)header[48])<<16 | ((size_t)header[49])<<24;
